requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
zeep>=4.2.1
pydantic>=2.9.0
python-dotenv>=1.0.0

